rm all the magic with Param replacing and unit deepcopies
It was necesary step for maintainability in future.
Initialization of Units and Interfaces is not divided in configuration, declaration and implementation.
This allows write much cleaner code and significantly speedup key parts of system.

Currently only unclear thing in project is .name vs ._name and ._names on Param.
Due case sensitivity/nosensivity, interface names and back references this is maybe
best achiavable solution. This is internal thing and it does not disturb
those who works with this lib.